requirements.txt
IometePyHive
thrift==0.13.0
sasl==0.2.1
thrift-sasl==0.4.2